+++ sdk/identity/azure-identity/azure/identity/_internal/__init__.py
@@ -4,14 +4,13 @@
 # ------------------------------------
 import os
 import logging
-from typing import TYPE_CHECKING
+from typing import List
+from azure.core.exceptions import ClientAuthenticationError
 
 from six.moves.urllib_parse import urlparse
 
 from .._constants import EnvironmentVariables, KnownAuthorities
 
-if TYPE_CHECKING:
-    from typing import Any, Optional
 
 try:
     from contextvars import ContextVar
@@ -66,11 +65,17 @@ def validate_tenant_id(tenant_id):
         )
 
 
-def resolve_tenant(default_tenant, tenant_id=None, **_):
-    # type: (str, Optional[str], **Any) -> str
+def resolve_tenant(
+        default_tenant: str,
+        tenant_id: str = None,
+        *,
+        additionally_allowed_tenant_ids: List[str] = [],
+        **_) -> str:
     """Returns the correct tenant for a token request given a credential's configuration"""
-    if tenant_id is None:
+    if tenant_id is None or tenant_id == default_tenant:
         return default_tenant
+    if not default_tenant:
+        return tenant_id
     if (
         default_tenant == "adfs"
         or os.environ.get(EnvironmentVariables.AZURE_IDENTITY_DISABLE_MULTITENANTAUTH)
@@ -79,10 +84,18 @@ def resolve_tenant(default_tenant, tenant_id=None, **_):
                      "but the configured value was used since multi tenant authentication has been disabled. "
                      "Configured tenant ID: %s, Requested tenant ID %s", default_tenant, tenant_id)
         return default_tenant
-    _LOGGER.info("A token was requested for a different tenant than was configured on the credential, "
-                 "and the requested tenant ID was used to authenticate. Configured tenant ID: %s, "
-                 "Requested tenant ID %s", default_tenant, tenant_id)
-    return tenant_id
+    if additionally_allowed_tenant_ids == ['*'] or tenant_id in additionally_allowed_tenant_ids:
+        _LOGGER.info("A token was requested for a different tenant than was configured on the credential, "
+                     "and the requested tenant ID was used to authenticate. Configured tenant ID: %s, "
+                     "Requested tenant ID %s", default_tenant, tenant_id)
+        return tenant_id
+    raise ClientAuthenticationError(
+                message='The current credential is not configured to acquire tokens for tenant {}. '
+                        'To enable acquiring tokens for this tenant add it to the additionally_allowed_tenant_ids '
+                        'when creating the credential, or add "*" to additionally_allowed_tenant_ids to allow '
+                        'acquiring tokens for any tenant.'.format(tenant_id)
+            )
+
 
 
 # pylint:disable=wrong-import-position