+++ sdk/identity/azure-identity/azure/identity/_internal/msal_credentials.py
@@ -3,6 +3,7 @@
 # Licensed under the MIT License.
 # ------------------------------------
 import os
+from typing import TYPE_CHECKING, List, Union, Dict
 
 import msal
 
@@ -18,14 +19,20 @@ except ImportError:
 
 if TYPE_CHECKING:
     # pylint:disable=ungrouped-imports,unused-import
-    from typing import Any, Dict, Optional, Union
+    from typing import Any
 
 
 class MsalCredential(object):
     """Base class for credentials wrapping MSAL applications"""
 
-    def __init__(self, client_id, client_credential=None, **kwargs):
-        # type: (str, Optional[Union[str, Dict]], **Any) -> None
+    def __init__(
+            self,
+            client_id: str,
+            client_credential: Union[str, Dict] = None,
+            *,
+            additionally_allowed_tenant_ids: List[str] = [],
+            **kwargs
+    ) -> None:
         authority = kwargs.pop("authority", None)
         # self._validate_authority = kwargs.pop("validate_authority", True)
         self._authority = normalize_authority(authority) if authority else get_default_authority()
@@ -36,6 +43,7 @@ class MsalCredential(object):
         self._client_applications = {}  # type: Dict[str, msal.ClientApplication]
         self._client_credential = client_credential
         self._client_id = client_id
+        self._additionally_allowed_tenant_ids = additionally_allowed_tenant_ids
 
         self._cache = kwargs.pop("_cache", None)
         if not self._cache:
@@ -60,7 +68,11 @@ class MsalCredential(object):
 
     def _get_app(self, **kwargs):
         # type: (**Any) -> msal.ClientApplication
-        tenant_id = resolve_tenant(self._tenant_id, **kwargs)
+        tenant_id = resolve_tenant(
+            self._tenant_id,
+            additionally_allowed_tenant_ids=self._additionally_allowed_tenant_ids,
+            **kwargs
+        )
         if tenant_id not in self._client_applications:
             # CP1 = can handle claims challenges (CAE)
             capabilities = None if "AZURE_IDENTITY_DISABLE_CP1" in os.environ else ["CP1"]