+++ sdk/identity/azure-identity/tests/test_device_code_credential.py
@@ -263,7 +263,7 @@ def test_tenant_id():
 
     callback = Mock()
     credential = DeviceCodeCredential(
-        client_id=client_id, prompt_callback=callback, transport=transport, instance_discovery=False,
+        client_id=client_id, prompt_callback=callback, transport=transport, instance_discovery=False, additionally_allowed_tenant_ids=['*']
     )
 
     now = datetime.datetime.utcnow()