+++ sdk/identity/azure-identity/tests/test_vscode_credential_async.py
@@ -282,7 +282,7 @@ async def test_multitenant_authentication():
         return mock_response(json_payload=build_aad_response(access_token=token))
 
     credential = get_credential(
-        tenant_id=first_tenant, transport=mock.Mock(send=send)
+        tenant_id=first_tenant, transport=mock.Mock(send=send), additionally_allowed_tenant_ids=['*']
     )
     with mock.patch(GET_REFRESH_TOKEN, lambda _: "**"):
         token = await credential.get_token("scope")
@@ -310,7 +310,7 @@ async def test_multitenant_authentication_not_allowed():
         token = expected_token if tenant == expected_tenant else expected_token * 2
         return mock_response(json_payload=build_aad_response(access_token=token))
 
-    credential = get_credential(tenant_id=expected_tenant, transport=mock.Mock(send=send))
+    credential = get_credential(tenant_id=expected_tenant, transport=mock.Mock(send=send), additionally_allowed_tenant_ids=['*'])
 
     with mock.patch(GET_REFRESH_TOKEN, lambda _: "**"):
         token = await credential.get_token("scope")