+++ sdk/identity/azure-identity/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Features Added
 
+- Added Windows Web Account Manager (WAM) Brokered Authentication Support. ([#23687](https://github.com/Azure/azure-sdk-for-python/issues/23687))
+
 ### Breaking Changes
 
 ### Bugs Fixed